+++ src/server/crowi/index.js
@@ -333,9 +333,10 @@ Crowi.prototype.setupSearcher = async function() {
   const SearchService = require('@server/service/search');
   const searchService = new SearchService(this);
 
-  if (searchService.isAvailable) {
+  if (searchService.isConfigured) {
     try {
-      this.searcher = searchService;
+      this.searchService = searchService;
+      this.searcher = searchService; // TODO: use `searchService` instead of `searcher`
     }
     catch (e) {
       logger.error('Error on setup searcher', e);