+++ src/server/routes/apiv3/index.js
@@ -35,5 +35,7 @@ module.exports = (crowi) => {
 
   router.use('/statistics', require('./statistics')(crowi));
 
+  router.use('/search', require('./search')(crowi));
+
   return router;
 };