+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## v3.6.5-RC
 
 * Impromvement: Add `checkMiddlewaresStrictly` query option to Healthcheck API
+* Support: Upgrade libs
+    * react-bootstrap-typeahead
 
 ## v3.6.4